+++ easy_orm_postgres/lib/easy_orm_postgres.dart
@@ -1,0 +1,3 @@
+export 'package:postgres/postgres.dart';
+export 'package:freezed_annotation/freezed_annotation.dart';
+export 'package:json_annotation/json_annotation.dart';